requirements.txt
openai
python-dotenv
flask
chromadb
fastapi
pydantic-settings